langchain1.0_basic/pyproject.toml
[project]
name = "langchain-a2a-mcp-example"
version = "0.1.0"
description = "LangChain 1.0 example with Ollama, MCP, and A2A integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-community>=0.3.0",
    "ollama>=0.4.0",
    "uvicorn>=0.34.0",
    "httpx>=0.28.1",
    "pydantic>=2.10.0",
    "fastmcp>=0.7.0",
    "a2a-sdk>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["agents", "utils", "async_examples", "examples"]
